backend/requirements.txt
# Flask Framework
Flask==3.1.2
Flask-CORS==6.0.1
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.7.1
Flask-Limiter==3.10.0

# Database
SQLAlchemy==2.0.44

# API Requests
requests==2.32.5

# Phone Number Processing
phonenumbers==8.13.50

# Environment Variables
python-dotenv==1.0.0

# Web Scraping (optional)
beautifulsoup4==4.14.2

# Task Queue
celery==5.3.1
redis==4.6.0

# Environment Variables
python-dotenv==1.0.0

# Validation
marshmallow==3.20.1
email-validator==2.0.0

# Security
cryptography==41.0.3
bcrypt==4.0.1

# Rate Limiting
Flask-Limiter==3.5.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# Testing
pytest==7.4.0
pytest-flask==1.2.0
pytest-cov==4.1.0

# Code Quality
flake8==6.1.0
black==23.7.0
